third_party/WebKit/LayoutTests/fast/mediastream/RTCPeerConnection-createOffer-expected.txt

Tests RTCPeerConnection createOffer.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testExecutionOrderClosedConnection() did not throw exception.
PASS pc.createOffer(createOfferSucceeded1, unexpectedCallback, {voiceActivityDetection:true, iceRestart:true}); did not throw exception.
PASS error.name is "InvalidStateError"
PASS error.toString() is "InvalidStateError: The RTCPeerConnection's signalingState is 'closed'."
PASS events is [1,2,3]
PASS createOfferSucceeded1 was called.
PASS sessionDescription.type is "offer"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed1); did not throw exception.
PASS expectedCreateOfferFailed1 was called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed2); did not throw exception.
PASS expectedCreateOfferFailed2 was called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed3, {}); did not throw exception.
PASS expectedCreateOfferFailed3 was called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed4, {voiceActivityDetection:false}); did not throw exception.
PASS expectedCreateOfferFailed4 was called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(createOfferSucceeded2, unexpectedCallback, {iceRestart:true}); did not throw exception.
PASS createOfferSucceeded2 was called.
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed5, {offerToReceiveVideo:1, offerToReceiveAudio:0, voiceActivityDetection:false, iceRestart:true}); did not throw exception.
PASS expectedCreateOfferFailed5 was called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed6, {offerToReceiveVideo:-1, offerToReceiveAudio:0}); did not throw exception.
PASS expectedCreateOfferFailed6 called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS pc.createOffer(unexpectedCallback, expectedCreateOfferFailed7, {offerToReceiveVideo:0, offerToReceiveAudio:-1}); did not throw exception.
PASS expectedCreateOfferFailed7 called.
PASS error.name is "OperationError"
PASS error.toString() is "OperationError: TEST_ERROR"
PASS successfullyParsed is true

TEST COMPLETE
